services/orchestrator/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
psycopg2-binary==2.9.9
boto3==1.42.75
pydantic==2.12.5